+++ source/Bookmark.h
@@ -4,6 +4,7 @@
 
 #include "TextSelection.h"
 #include <QChar>
+#include <cstdint>
 
 // max. # of bookmarks (one per letter & #)
 constexpr int MAX_MARKS = 36;
@@ -11,7 +12,7 @@ constexpr int MAX_MARKS = 36;
 // Element in bookmark table
 struct Bookmark {
     QChar         label;
-    int           cursorPos;
+    int64_t       cursorPos;
 	TextSelection sel;
 };
