+++ source/UndoInfo.cpp
@@ -1,5 +1,5 @@
 
 #include "UndoInfo.h"
 
-UndoInfo::UndoInfo(UndoTypes undoType, int start, int end) : type(undoType), startPos(start), endPos(end) {
+UndoInfo::UndoInfo(UndoTypes undoType, int64_t start, int64_t end) : type(undoType), startPos(start), endPos(end) {
 }
